refactor: Extract AccessRightSection's modal content to a new component
