+++ src/components/Permissions/AccessRightsSection.jsx
@@ -19,7 +19,7 @@ import {
 import withAllLocales from '../../lib/withAllLocales'
 import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
 import { Dialog } from 'cozy-ui/transpiled/react/CozyDialogs'
-import Alert from 'cozy-ui/transpiled/react/Alert'
+import AccessRightsDetailsSection from './AccessRightsDetailsSection'
 
 const AccessRightsSection = ({
   sortedPermissionsByName,
@@ -37,15 +37,6 @@ const AccessRightsSection = ({
   }
 
   const closeModal = () => setModalOpened(false)
-
-  const displayedPermissions = modalData && displayPermissions(modalData.verbs)
-
-  const hasReadPermissions =
-    displayedPermissions === 'Permissions.read' ||
-    displayedPermissions === 'Permissions.readAndWrite'
-  const hasWritePermissions =
-    displayedPermissions === 'Permissions.write' ||
-    displayedPermissions === 'Permissions.readAndWrite'
 
   if (!sortedPermissionsByName || sortedPermissionsByName.length < 1) {
     return
@@ -129,60 +120,7 @@ const AccessRightsSection = ({
           onClose={closeModal}
           title={modalData?.title ? modalData.title : ''}
           content={
-            <>
-              <Typography variant="h5" className="">
-                {modalData
-                  ? modalData.isRemoteDoctypes
-                    ? t('Permissions.exit_right', {
-                        app: slugName.toUpperCase(),
-                        doctype: modalData.title.toLowerCase()
-                      })
-                    : t('Permissions.access_right', {
-                        app: slugName.toUpperCase(),
-                        doctype: modalData.title.toLowerCase()
-                      })
-                  : ''}
-              </Typography>
-              <Alert
-                severity={modalData?.isRemoteDoctypes ? 'warning' : 'success'}
-                className="u-mt-half"
-              >
-                {modalData?.isRemoteDoctypes
-                  ? t('Permissions.exit_right_description')
-                  : t('Permissions.access_right_description', {
-                      app: slugName.toUpperCase()
-                    })}
-              </Alert>
-              <Typography variant="h5" className="u-mt-1-half">
-                {t('Permissions.right_reason', { app: slugName.toUpperCase() })}
-              </Typography>
-              <Typography className="u-mt-half">
-                {modalData?.description}
-              </Typography>
-              <Typography variant="h5" className="u-mt-1-half">
-                {t('Permissions.details')}
-              </Typography>
-              {hasReadPermissions && (
-                <>
-                  <Typography className="u-mt-half">
-                    {t('Permissions.read_right_title')}
-                  </Typography>
-                  <Typography variant="caption">
-                    {t('Permissions.read_right_text')}
-                  </Typography>
-                </>
-              )}
-              {hasWritePermissions && (
-                <>
-                  <Typography className="u-mt-half">
-                    {t('Permissions.write_right_title')}
-                  </Typography>
-                  <Typography variant="caption">
-                    {t('Permissions.write_right_text')}
-                  </Typography>
-                </>
-              )}
-            </>
+            <AccessRightsDetailsSection modalData={modalData} slug={slugName} />
           }
         />
       )}
